# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Guitar learning app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teacher/classes:
    get:
      operationId: listMyClasses
      tags: [teacher]
      summary: List classes owned by the current teacher
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassWithStats"
    post:
      operationId: createClass
      tags: [teacher]
      summary: Create a new class
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassWithStats"
  /teacher/classes/{id}/expand:
    post:
      operationId: expandClassCapacity
      tags: [teacher]
      summary: Add additional student slots (and codes) to an existing class
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpandClassBody"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassWithStats"
        "400":
          description: Limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /teacher/classes/{id}/unlock-next:
    post:
      operationId: unlockNextLevel
      tags: [teacher]
      summary: Unlock the next level for a class
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassWithStats"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CreateClassBody:
      type: object
      required: [name, studentCount]
      properties:
        name:
          type: string
          minLength: 1
        studentCount:
          type: integer
          minimum: 1
    ExpandClassBody:
      type: object
      required: [additional]
      properties:
        additional:
          type: integer
          minimum: 1
    StudentCodeInfo:
      type: object
      required: [code, used]
      properties:
        code:
          type: string
        used:
          type: boolean
        usedByName:
          type: string
          nullable: true
    ClassWithStats:
      type: object
      required: [id, name, levelUnlocked, studentCount, studentCapacity, usedStudentCount, remainingSlots, unusedStudentCodes, studentCodes]
      properties:
        id:
          type: string
        name:
          type: string
        levelUnlocked:
          type: integer
        studentCount:
          type: integer
        studentCapacity:
          type: integer
        usedStudentCount:
          type: integer
        remainingSlots:
          type: integer
        unusedStudentCodes:
          type: integer
        studentCodes:
          type: array
          items:
            $ref: "#/components/schemas/StudentCodeInfo"
...