# lib/api-zod/src/generated/types/expandClassBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guitar learning app API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpandClassBody {
  /** @minimum 1 */
  additional: number;
}