# lib/api-zod/src/generated/types/studentCodeInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guitar learning app API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentCodeInfo {
  code: string;
  used: boolean;
  usedByName?: string | null;
}